+++ src/textview/tree.rs
@@ -11,14 +11,14 @@ use crate::painter::Painter;
 use super::view::TextView;
 
 pub(crate) struct TextTree {
-    border_width: f32,
+    border_width: u32,
     root: Node,
 }
 
 impl TextTree {
     pub(crate) fn new(
-        rect: Rect<f32, PixelSize>,
-        border_width: f32,
+        rect: Rect<u32, PixelSize>,
+        border_width: u32,
         buffer: BufferHandle,
         view_id: BufferViewId,
     ) -> TextTree {
@@ -28,7 +28,7 @@ impl TextTree {
         }
     }
 
-    pub(crate) fn set_rect(&mut self, rect: Rect<f32, PixelSize>) {
+    pub(crate) fn set_rect(&mut self, rect: Rect<u32, PixelSize>) {
         self.root.set_rect(rect)
     }
 
@@ -40,7 +40,7 @@ impl TextTree {
         self.root.active_mut()
     }
 
-    pub(crate) fn move_cursor_to_point(&mut self, point: Point2D<f32, PixelSize>) {
+    pub(crate) fn move_cursor_to_point(&mut self, point: Point2D<i32, PixelSize>) {
         self.root.move_cursor_to_point(point);
     }
 
@@ -54,19 +54,19 @@ impl TextTree {
 }
 
 struct Node {
-    rect: Rect<f32, PixelSize>,
+    rect: Rect<u32, PixelSize>,
     view: TextView,
 }
 
 impl Node {
-    fn new_leaf(rect: Rect<f32, PixelSize>, buffer: BufferHandle, view_id: BufferViewId) -> Node {
+    fn new_leaf(rect: Rect<u32, PixelSize>, buffer: BufferHandle, view_id: BufferViewId) -> Node {
         Node {
             rect,
             view: TextView::new(rect, buffer, view_id),
         }
     }
 
-    fn set_rect(&mut self, rect: Rect<f32, PixelSize>) {
+    fn set_rect(&mut self, rect: Rect<u32, PixelSize>) {
         self.rect = rect;
         self.view.set_rect(rect)
     }
@@ -79,8 +79,8 @@ impl Node {
         &mut self.view
     }
 
-    fn move_cursor_to_point(&mut self, point: Point2D<f32, PixelSize>) {
-        if !self.rect.contains(point) {
+    fn move_cursor_to_point(&mut self, point: Point2D<i32, PixelSize>) {
+        if !self.rect.cast().contains(point) {
             return;
         }
         self.view.move_cursor_to_point(point);
